24/08/26 중간 커밋

+++ src/main/java/org/zerock/chain/pse/dto/NotificationDTO.java
@@ -5,7 +5,8 @@ import java.time.LocalDateTime;
 public class NotificationDTO {
 
     private long notificationNo;
-    private Integer empNo;
+    private Long empNo;
+
     private String notificationType;
     private long referenceId;
     private String  notificationMessage;
